Rename "internal/protocol" to "internal/web".
