+++ internal/web/cookie.go
@@ -1,4 +1,4 @@
-package protocol
+package web
 
 import (
 	"net/http"
