+++ internal/web/default.go
@@ -1,4 +1,4 @@
-package protocol
+package web
 
 import (
 	"crypto/tls"
